Move SearchNode into the definition of the class that uses it
