+++ src/Engine/Utils.java
@@ -155,52 +155,6 @@ public class Utils
   }
 
   /**
-   * Utility class used for pathfinding. Optionally holds a
-   *   reference to a parent node for path reconstruction.<p>
-   * Caveat emptor: the SearchNode quacks like an XYCoord for equality checks
-   */
-  public static class SearchNode extends XYCoord
-  {
-    private static final long serialVersionUID = 2637721435469761667L;
-    public SearchNode parent;
-    public HashSet<SearchNode> allParents;
-
-    public SearchNode(int x, int y)
-    {
-      this(x, y, null);
-    }
-
-    public SearchNode(XYCoord coord, SearchNode parent)
-    {
-      this(coord.x, coord.y, parent);
-    }
-    public SearchNode(int x, int y, SearchNode parent)
-    {
-      super(x, y);
-      this.parent = parent;
-    }
-    public XYCoord getCoordinates()
-    {
-      return this;
-    }
-    public GamePath getMyPath()
-    {
-      GamePath aPath = new GamePath();
-
-      SearchNode currentNode = this;
-      // Add all of the points on the route to our waypoint list.
-      while (currentNode != null)
-      {
-        // Since we're iterating dest->start, each point is the new "first" point.
-        aPath.addWaypoint(0, currentNode.x, currentNode.y);
-        currentNode = currentNode.parent;
-      }
-
-      return aPath;
-    }
-  }
-
-  /**
    * Returns a list of all vacant industries a army owns
    */
   public static ArrayList<XYCoord> findUsableProperties(Army army, GameMap map)
