+++ src/Test/TestUnitMovement.java
@@ -10,9 +10,9 @@ import Engine.GameAction;
 import Engine.GameInstance;
 import Engine.GameScenario;
 import Engine.PathCalcParams;
+import Engine.PathCalcParams.SearchNode;
 import Engine.GamePath;
 import Engine.Utils;
-import Engine.Utils.SearchNode;
 import Engine.XYCoord;
 import Engine.UnitActionLifecycles.WaitLifecycle;
 import Terrain.MapInfo;
